+++ client/src/components/threadspage/userthreads/UserThreads.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 
-export default function UserThreads(props, { handlereviewModalTog}) {
+export default function UserThreads(props, { handlereviewModalTog, onViewChange}) {
 
     console.log(typeof(props.date));
       const date = props.date.split(' ');
@@ -44,7 +44,7 @@ export default function UserThreads(props, { handlereviewModalTog}) {
             <p className='m-0 pt-1 pb-1'>(edit)</p>
             <button
               className='thread-open-btn'
-              onClick={() => handlereviewModalTog(props.key)}
+              onClick={() => onViewChange()}
             >
               Open
             </button>
